Added initial implementation of encryption scheme for password and more
Added initial implementation and initial accompanying tests for encryption scheme. Also updated UserSettings.java file to accommodate the new requirements, mainly adding columns for password, salt, and IV, the latter two of which are necessary for verification of password. The password will be used to store the API Key and Secret securely.